navigation redirect not working
